Refactor config file into a dedicated package
This commit moves the config file into a dedicated `config` package with functions for getting and setting OpenAI config variables. The `getConfig` function now returns an error if the `OpenAIKey` field is empty or set to the default value of "your_api_key". Additionally, a `config_test` file has been added to test the `getConfigLocation` and `SetConfig` functions.